comp fix in func test

+++ tasks/levonychev_i_radix_batcher_sort/tests/functional/main.cpp
@@ -1,6 +1,7 @@
 #include <gtest/gtest.h>
 
 #include <array>
+#include <cstddef>
 #include <string>
 #include <tuple>
 
@@ -32,7 +33,7 @@ class LevonychevIRadixBatcherSortRunFuncTestsThreads : public ppc::util::BaseRun
   }
 
   bool CheckTestOutputData(OutType &output_data) final {
-    for (int i = 1; i < output_data.size(); ++i) {
+    for (size_t i = 1; i < output_data.size(); ++i) {
       if (output_data[i - 1] > output_data[i]) {
         return false;
       }
